Rename some modules for consistant naming patterns
