+++ xaitk_saliency/__init__.py
@@ -1,8 +1,9 @@
 import pkg_resources
 
 from .interfaces.gen_classifier_conf_sal import GenerateClassifierConfidenceSaliency  # noqa: F401
+from .interfaces.gen_descriptor_sim_sal import GenerateDescriptorSimilaritySaliency  # noqa: F401
+from .interfaces.gen_detector_prop_sal import GenerateDetectorProposalSaliency  # noqa: F401
 from .interfaces.gen_image_classifier_blackbox_sal import GenerateImageClassifierBlackboxSaliency  # noqa: F401
-from .interfaces.gen_similarity_sal import GenerateDescriptorSimilaritySaliency  # noqa: F401
 from .interfaces.perturb_image import PerturbImage  # noqa: F401
